OXDEV-3696 Fix string issue for getCurrentShopId in context

+++ source/Internal/Transition/Utility/Context.php
@@ -56,7 +56,7 @@ class Context extends BasicContext implements ContextInterface
      */
     public function getCurrentShopId(): int
     {
-        return Registry::getConfig()->getShopId();
+        return (int)Registry::getConfig()->getShopId();
     }
 
     /**
